refactor client page, added delete client flow, fixed purchase

+++ src/assets/js/controller/ClientDeleteController.js
@@ -2,13 +2,16 @@ APP.controller('ClientDeleteController', ['$scope', '$rootScope', '$state', '$st
     function ($scope, $rootScope, $state, $stateParams, ClientService) {
 
         $scope.delete = function (client) {
-            ClientService.delete(client).then(function (deleteSuccess) {
-                $state.go('clientList');
-            })
-        }
-
-        $scope.back = function (client) {
-            $state.go('clientList');
+            if ($scope.purchasesCount > 0) {
+                alert(client.name + ' ' + client.surname + ' ha acquistato ' + $scope.purchasesCount + ' prodotti. Non è possibile rimuoverlo!')
+            } else {
+                ClientService.delete(client).then(function (deleteSuccess) {
+                    UIkit.modal('#client-delete').hide();
+                    $state.go('clientsList');
+                }, function (error) {
+                    alert(error.data.message);
+                })
+            }
         }
 
     }]);
